# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API REST de Lebane para gestionar departamentos en venta: alta, edición, listado con filtros, fotos y consultas de interesados.\n\n**Convenciones**\n\n- **Errores**: toda respuesta 4xx/5xx tiene el mismo esquema `ApiError`, con un código estable en `error`, un mensaje apto para mostrar al usuario y el `requestId` para soporte. Nunca incluye stack traces, SQL ni detalles de infraestructura. Los errores de validación detallan cada campo en `fieldErrors` (las claves son las rutas del body o los nombres de los parámetros, p. ej. `direccion.ciudad`).\n- **Correlation ID**: cada request puede enviar `X-Request-Id`; si no lo envía o no es válido, el backend genera uno. La respuesta siempre lo devuelve y queda en todos los logs del request.\n- **Concurrencia optimista**: el detalle, el alta y la edición devuelven `ETag` con la versión del departamento. Enviarlo en `If-Match` al editar evita pisar cambios de otro usuario (412 si cambió).\n- **Paginación**: `page` (desde 0) y `size` (1 a 100), con una ventana máxima de 10.000 resultados por búsqueda. La respuesta usa el formato estándar de Spring Data (`content` + `page`).\n- **Disponibilidad**: si una dependencia no crítica falla (storage de imágenes, proveedor de direcciones), solo se degradan las operaciones que la usan: 503 `STORAGE_UNAVAILABLE` al subir fotos, respuesta `degradado: true` en el autocompletado. El resto de la API sigue funcionando.\n",
...
    "version" : "v1"
...
  "servers" : [ {
...
  "paths" : {
    "/api/v1/departamentos" : {
      "get" : {
        "description" : "Listado paginado con filtros y orden, resueltos en la base de datos (nunca en memoria). Cada ítem trae su foto principal y la cantidad de fotos y de consultas. Los filtros se combinan con AND; los valores de `estado` se combinan con OR.",
        "operationId" : "listar",
        "parameters" : [ {
          "description" : "Texto a buscar en el título (sin distinguir mayúsculas; mínimo 3 caracteres)",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Texto a buscar en el título (sin distinguir mayúsculas; mínimo 3 caracteres)",
            "maxLength" : 100,
            "minLength" : 3
          }
        }, {
          "description" : "Ciudad exacta (sin distinguir mayúsculas)",
          "in" : "query",
          "name" : "ciudad",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Ciudad exacta (sin distinguir mayúsculas)",
            "maxLength" : 80,
            "minLength" : 0
          }
        }, {
          "description" : "Estados a incluir; repetir el parámetro para varios (estado=DISPONIBLE&estado=RESERVADO)",
          "in" : "query",
          "name" : "estado",
          "required" : false,
          "schema" : {
            "type" : "array",
            "description" : "Estados a incluir; repetir el parámetro para varios (estado=DISPONIBLE&estado=RESERVADO)",
            "items" : {
              "type" : "string",
              "description" : "Estado comercial. Transiciones: DISPONIBLE ⇄ RESERVADO y ambos → VENDIDO. VENDIDO es final: no se edita, no cambia sus fotos ni recibe consultas, y no se puede publicar directamente como vendido.",
              "enum" : [ "DISPONIBLE", "RESERVADO", "VENDIDO" ]
            }
          }
        }, {
          "description" : "Moneda. Obligatoria si se filtra por precio (ARS y USD no son comparables)",
          "in" : "query",
          "name" : "moneda",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Moneda del precio: pesos argentinos o dólares.",
            "enum" : [ "ARS", "USD" ]
          }
        }, {
          "description" : "Precio mínimo (requiere moneda)",
          "in" : "query",
          "name" : "precioMin",
          "required" : false,
          "schema" : {
            "type" : "number",
            "description" : "Precio mínimo (requiere moneda)",
            "minimum" : 0
          }
        }, {
          "description" : "Precio máximo (requiere moneda; mayor o igual que precioMin)",
          "in" : "query",
          "name" : "precioMax",
          "required" : false,
          "schema" : {
            "type" : "number",
            "description" : "Precio máximo (requiere moneda; mayor o igual que precioMin)",
            "minimum" : 0
          }
        }, {
          "description" : "Ambientes mínimos",
          "in" : "query",
          "name" : "ambientesMin",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Ambientes mínimos",
            "maximum" : 20,
            "minimum" : 1
          }
        }, {
          "description" : "Dormitorios mínimos",
          "in" : "query",
          "name" : "dormitoriosMin",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Dormitorios mínimos",
            "maximum" : 19,
            "minimum" : 0
          }
        }, {
          "description" : "Baños mínimos",
          "in" : "query",
          "name" : "banosMin",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Baños mínimos",
            "maximum" : 10,
            "minimum" : 1
          }
        }, {
          "description" : "Superficie mínima en m²",
          "in" : "query",
          "name" : "superficieMin",
          "required" : false,
          "schema" : {
            "type" : "number",
            "description" : "Superficie mínima en m²",
            "minimum" : 0
          }
        }, {
          "description" : "Superficie máxima en m² (mayor o igual que superficieMin)",
          "in" : "query",
          "name" : "superficieMax",
          "required" : false,
          "schema" : {
            "type" : "number",
            "description" : "Superficie máxima en m² (mayor o igual que superficieMin)",
            "minimum" : 0
          }
        }, {
          "description" : "true: solo con fotos; false: solo sin fotos; omitido: todos",
          "in" : "query",
          "name" : "conImagenes",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "description" : "true: solo con fotos; false: solo sin fotos; omitido: todos"
          }
        }, {
          "description" : "Página, desde 0. (page + 1) × size no puede superar 10.000",
          "example" : 0,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Página, desde 0. (page + 1) × size no puede superar 10.000",
            "example" : 0,
            "minimum" : 0
          }
        }, {
          "description" : "Tamaño de página",
          "example" : 20,
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "description" : "Tamaño de página",
            "example" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        }, {
          "description" : "Orden: createdAt, precio o superficieM2, con ,asc o ,desc. El precio se ordena dentro de cada moneda; siempre se desempata por id",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt,desc",
            "description" : "Orden: createdAt, precio o superficieM2, con ,asc o ,desc. El precio se ordena dentro de cada moneda; siempre se desempata por id",
            "pattern" : "^(createdAt|precio|superficieM2)(,(asc|desc|ASC|DESC))?$"
          }
        }, {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelDepartamentoListItemResponse"
                }
              }
            },
            "description" : "Página de resultados (puede estar vacía)",
            "headers" : {
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Listar departamentos",
        "tags" : [ "Departamentos" ]
      },
      "post" : {
        "description" : "Da de alta un departamento. El backend genera el `codigo` comercial; si no se indica `estado`, queda DISPONIBLE. No se puede publicar directamente como VENDIDO (409 `TRANSICION_DE_ESTADO_INVALIDA`). Las fotos se suben después, de a una, con el endpoint de imágenes.",
        "operationId" : "crear",
        "parameters" : [ {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepartamentoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepartamentoDetailResponse"
                }
              }
            },
            "description" : "Departamento creado",
            "headers" : {
              "ETag" : {
                "description" : "Versión del departamento (p. ej. \"3\"). Enviarla en If-Match al editar para no pisar cambios de otro usuario.",
                "schema" : {
                  "type" : "string",
                  "example" : "\"0\""
                },
                "style" : "simple"
              },
              "Location" : {
                "description" : "Ruta relativa del departamento creado",
                "schema" : {
                  "type" : "string",
                  "example" : "/api/v1/departamentos/1"
                },
                "style" : "simple"
              },
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Crear un departamento",
        "tags" : [ "Departamentos" ]
      }
    },
...
    "/api/v1/departamentos/{id}" : {
      "get" : {
        "description" : "Detalle completo: dirección, fotos ordenadas y cantidad de consultas.",
        "operationId" : "obtener",
        "parameters" : [ {
          "description" : "ID del departamento",
          "example" : 1,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepartamentoDetailResponse"
                }
              }
            },
            "description" : "Departamento encontrado",
            "headers" : {
              "ETag" : {
                "description" : "Versión del departamento (p. ej. \"3\"). Enviarla en If-Match al editar para no pisar cambios de otro usuario.",
                "schema" : {
                  "type" : "string",
                  "example" : "\"3\""
                },
                "style" : "simple"
              },
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Obtener un departamento",
        "tags" : [ "Departamentos" ]
      },
      "put" : {
        "description" : "Reemplazo completo de los datos (no modifica fotos ni consultas). Enviar en `If-Match` el ETag obtenido al leerlo: si otro usuario lo modificó desde entonces, responde 412 en lugar de pisar sus cambios. Reglas de estado: DISPONIBLE ⇄ RESERVADO y ambos → VENDIDO; un departamento VENDIDO no se puede modificar (409 `DEPARTAMENTO_NO_DISPONIBLE`) y un cambio de estado no permitido responde 409 `TRANSICION_DE_ESTADO_INVALIDA`.",
        "operationId" : "actualizar",
        "parameters" : [ {
          "description" : "ID del departamento",
          "example" : 1,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "description" : "ETag leído (p. ej. \"3\"). Opcional: sin él, la edición no verifica la versión.",
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "\"3\""
          }
        }, {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepartamentoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepartamentoDetailResponse"
                }
              }
            },
            "description" : "Departamento actualizado",
            "headers" : {
              "ETag" : {
                "description" : "Nueva versión del departamento",
                "schema" : {
                  "type" : "string",
                  "example" : "\"4\""
                },
                "style" : "simple"
              },
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "412" : {
            "$ref" : "#/components/responses/PreconditionFailed"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Editar un departamento",
        "tags" : [ "Departamentos" ]
      }
    },
    "/api/v1/departamentos/{id}/consultas" : {
      "post" : {
        "description" : "Registra la consulta de un interesado. Un departamento VENDIDO no recibe consultas (409 `DEPARTAMENTO_NO_DISPONIBLE`). La respuesta no devuelve los datos personales enviados.",
        "operationId" : "crearConsulta",
        "parameters" : [ {
          "description" : "ID del departamento",
          "example" : 1,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsultaRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsultaCreatedResponse"
                }
              }
            },
            "description" : "Consulta registrada",
            "headers" : {
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Enviar una consulta",
        "tags" : [ "Consultas" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "json" : {
                "summary" : "JSON mal formado",
                "value" : {
                  "error" : "BAD_REQUEST",
                  "message" : "El cuerpo de la solicitud no es JSON válido",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 400,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "validacion" : {
                "summary" : "Validación de campos",
                "value" : {
                  "error" : "VALIDATION_ERROR",
                  "fieldErrors" : {
                    "dormitorios" : "debe ser menor que la cantidad de ambientes",
                    "titulo" : "no debe estar vacío"
                  },
                  "message" : "La solicitud contiene datos inválidos",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 400,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Datos inválidos: body, parámetros o archivo. Cada campo con error se detalla en `fieldErrors`.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
...
      "InternalError" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "inesperado" : {
                "summary" : "Error inesperado",
                "value" : {
                  "error" : "INTERNAL_ERROR",
                  "message" : "Ocurrió un error inesperado; si persiste, informá el requestId",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 500,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Error inesperado. El mensaje es genérico; el `requestId` permite encontrar el detalle en los logs.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "schemas" : {
...
      "PagedModelDepartamentoListItemResponse" : {
        "type" : "object",
        "description" : "Página de resultados (formato estándar de Spring Data)",
        "properties" : {
          "content" : {
            "type" : "array",
            "description" : "Resultados de la página, en el orden pedido",
            "items" : {
              "$ref" : "#/components/schemas/DepartamentoListItemResponse"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
...
    }
  }
}